try to make travis OSX builds happy

+++ src/algorithms/approx_path_distance.cpp
@@ -8,7 +8,7 @@ using namespace std;
 size_t min_approx_path_distance(const PathPositionHandleGraph* graph, const pos_t& pos1, const pos_t& pos2, uint64_t max_search) {
     auto nearest1 = nearest_offsets_in_paths(graph, pos1, max_search);
     auto nearest2 = nearest_offsets_in_paths(graph, pos2, max_search);
-    size_t min_distance = numeric_limits<size_t>::max();
+    uint64_t min_distance = numeric_limits<uint64_t>::max();
     for (auto& p : nearest1) {
         auto q = nearest2.find(p.first);
         if (q != nearest2.end()) {
@@ -21,7 +21,7 @@ size_t min_approx_path_distance(const PathPositionHandleGraph* graph, const pos_
             }
         }
     }
-    return min_distance;
+    return (size_t)min_distance;
 }
     
 }
